Remade prototype level, added stone material.
The first prototype level had a scale that was way too big, so I remade it and applied a stone texture. The wall is made with ProBuilder, and it's tiling the texture weirdly - need to fix that. I also finished the MovingPlatform script., which uses Raycasting to make an object bounce between two targets.